:art: move auto add/delete functions into own helpers file
