+++ server/src/helpers/index.js
@@ -1,5 +1,6 @@
 const slug = require('slugify')
 
+const { questionDeleteCertif, answerUpdateCertif } = require('./certified')
 const diffTags = require('./diffTags')
 const emojify = require('./emojify')
 const history = require('./history')
@@ -12,6 +13,8 @@ const slugify = s => slug(s).toLowerCase()
 
 module.exports = {
   ctxUser,
+  questionDeleteCertif,
+  answerUpdateCertif,
   diffTags,
   emojify,
   history,
